+++ src/hooks/index.ts
@@ -3,5 +3,6 @@ export * from './use-me';
 export * from './use-color-scheme';
 export * from './use-theme-color';
 export * from './use-device-registration';
+export * from './use-complete-onboarding';
 export * from './auth/use-email-sign-in';
 export * from './auth/use-email-sign-up';